build(vite): Front build passed from cra to vite
Previously the project was builded with create-react-app but is now builded with vite.

During the developpement i have seen that the node_modules module was not indicated in the .gitignore file causing the push of multiples library files that should normaly be installed by the developpers independently with the command npm install. To correct that mistake i have deleted the node_module folder in this commit and added node_modules in the .gitignore file.